Update script that runs docker container to read env vars from a host-specific config file.

+++ run-docker.sh
@@ -1,8 +1,18 @@
 #!/bin/bash
 
-# When calling this, may want to provide 
+# Read configuration file, if it exists
+FILENAME=`basename $0`
+CONFIG=/etc/default/${FILENAME}
+[ -e ${CONFIG} ] && . ${CONFIG}
+
+SLIMSERVER_IMAGE_USERNAME=${DOCKER_SLIMSERVER_IMAGE_USERNAME:-$USER}
+
 SLIMSERVER_UID=${DOCKER_SLIMSERVER_UID:-1001}
 SLIMSERVER_GID=${DOCKER_SLIMSERVER_GID:-1001}
+
+DOCKER_RESTART=" --restart=unless-stopped "
+#DOCKER_RESTART=" --restart=no "
+SLIMSERVER_ARGS=${DOCKER_SLIMSERVER_ARGS:-$DOCKER_RESTART}
 
 SLIMSERVER_IP=${DOCKER_SLIMSERVER_IP}
 
@@ -10,8 +20,12 @@ SLIMSERVER_MEDIA_DIR=${DOCKER_SLIMSERVER_MEDIA_DIR:-/var/lib/squeezeboxserver/me
 
 SLIMSERVER_NAME=lms
 
-docker stop ${SLIMSERVER_NAME}
-docker rm ${SLIMSERVER_NAME}
+DOCKER_COUNT=`docker ps -a -f name=${SLIMSERVER_NAME} | grep -v IMAGE | wc | awk '{print($1)}'`
+
+if [ "${DOCKER_COUNT}" != "0" ]; then
+  docker stop ${SLIMSERVER_NAME}
+  docker rm ${SLIMSERVER_NAME}
+fi
 
 DOCKER_ARGS="
   -p ${SLIMSERVER_IP}9000:9000 
@@ -25,11 +39,11 @@ DOCKER_ARGS="
   -e SLIMSERVER_UID=$SLIMSERVER_UID 
   -e SLIMSERVER_GID=$SLIMSERVER_GID
   --name=$SLIMSERVER_NAME
-  $USER/slimserver"
+  ${SLIMSERVER_IMAGE_USERNAME}/slimserver"
 
-if [ $# != 0 ]; then
-  echo 'Running as daemon'
-  docker run -d $DOCKER_ARGS
+if [ $# == 0 ]; then
+  echo 'Running as daemon; --restart=unless-stopped will restart the container on reboot'
+  docker run -d ${SLIMSERVER_ARGS} $DOCKER_ARGS
 else
   echo 'Running interactively'
   docker run --entrypoint /bin/bash -it --rm $DOCKER_ARGS
